backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-SocketIO==5.3.6
Flask-Cors==4.0.1
Flask-Limiter==3.7.0
marshmallow==3.21.3
psycopg2-binary==2.9.9
python-dotenv==1.0.1
bcrypt==4.1.3
gunicorn==22.0.0
eventlet==0.36.1
greenlet==3.0.3
python-dateutil==2.9.0
